Bugfix: point color wasn't applied when point_idx = 0
The color of points wasn't being applied when they're all the way to the left. In that case, the default light/dark colors are being applied instead.

To reproduce:

1. set a point color (that isn't black/white)
2. reduce RGB, HSL, or alpha to 0
3. point color is being shown in black (or white)

The issue is that the for loops for the bars in UI:highlight range from 1-width. That means that point_idx never matches when it is 0.